When a user clicks on a the profile image, they are taken to the phone's camera roll so they can select an image to have as their profile picture. The image is properly stored in Firebase's storage and the link is properly updated under pictureURL in the specific user's database. Still need to link the picture to the imageView so it is displayed.
