+++ Yoga/Info.plist
@@ -46,5 +46,7 @@
 	</array>
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
+	<key>NSPhotoLibraryUsageDescription</key>
+	<string>Access to your Photo Library is needed in order for you to select an profile image.</string>
 </dict>
 </plist>
